pyproject.toml
[project]
name = "donor-db-clinic"
version = "0.1.0"
description = ""
authors = [
    {name = "AndreaTomas00",email = "andrea.tomas00@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "xmltodict (>=0.14.2,<0.15.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "loguru (==0.7.2)",
    "typer (>=0.15.1,<0.16.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "pdfplumber (>=0.11.5,<0.12.0)",
    "streamlit (>=1.43.0,<2.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "rich (>=13.9.4,<14.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "pylint (>=3.3.5,<4.0.0)"
]

[tool.poetry]
name = "pipeline"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include ="src"}]

[tool.pylint]
max-line-length = 120
disable = ["C0111", "C0103"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
